refactor: move findOptionalFieldPaths to feat/travel-document
