+++ client/src/feat/travel-document/utils.test.ts
@@ -1,0 +1,7 @@
+import { describe, it } from "vitest";
+
+describe("findOptionalFieldPaths", () => {
+	it.todo(
+		"finds the number of fields which are optional when form is finalized",
+	);
+});
